peel off TestCase  into its own file
